% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aquamaps.R
\name{get_AquaMaps}
\alias{get_AquaMaps}
\title{Code to reduce AquaMaps to our requested species and area.}
\usage{
get_AquaMaps(
  PlanUnits,
  MinD = 0,
  MaxD = 200,
  Direc = file.path("~", "SpatPlan_Data")
)
}
\arguments{
\item{PlanUnits}{Planning Units as an \code{sf} object}

\item{MinD}{The minimum depth for AquaMaps data}

\item{MaxD}{The maximum depth for AquaMaps data}

\item{Direc}{The directory where the MME data is being stored. If not specified, the default location is assumed.}
}
\value{
An \code{sf} object of AquaMaps data
}
\description{
Code to reduce AquaMaps to our requested species and area.
}
\examples{
\dontrun{
CoralSea <- offshoredatr::get_area("Coral Sea",
                                   query_type = "iho",
                                   mregions_column = "name") \%>\%
  sf::st_transform("ESRI:54009") \%>\%
  sf::st_make_valid()
Aust <- rnaturalearth::ne_countries(country = "Australia", returnclass = "sf") \%>\%
  sf::st_transform("ESRI:54009") \%>\%
  sf::st_make_valid()
PUs <- spatialplanr::splnr_get_PlanningUnits(CoralSea,
                                             Aust,
                                             CellArea = 10000,
                                             Shape = "hexagon")
ggplot2::ggplot() +
  ggplot2::geom_sf(data = Aust, colour = "blue", fill = NA) +
  ggplot2::geom_sf(data = PUs, colour = "black", fill = NA)

AQM <- get_AquaMaps(PUs, MinD = 0, MaxD = 200,
                    Direc = file.path("~", "Nextcloud", "MME1DATA-Q1215", "SpatialLayers"))

(gg <- spatialplanr::splnr_plot_featureNo(AQM, Aust))
}
}
