% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_landMask.R
\name{get_landMask}
\alias{get_landMask}
\title{Very simple function to return a global land-sea mask.}
\usage{
get_landMask(nrow = 180, ncol = 360, scale = "large")
}
\arguments{
\item{nrow}{Number of rows in global raster}

\item{ncol}{Number of columns in global raster}

\item{scale}{The scale of the data to be downloaded from rnaturalearth.}
}
\value{
A SpatRaster from \code{terra}
}
\description{
The function uses rnaturalearth for masking the land. It defaults to 1
degree but any resolution can be used by changing \code{nrow} and \code{ncol}.
}
\examples{
r <- get_landMask(nrow = 180, ncol = 360, scale = "large")
}
