% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{longhurst}
\alias{longhurst}
\title{Longhurst Provinces}
\format{
\code{longhurst}
A data frame with XXXX rows and XXX columns:
\describe{
}
}
\source{
"XXXXXX"
}
\usage{
longhurst
}
\description{
An sf dataframe of
}
\keyword{datasets}
