% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fishcatch.R
\name{match_FishCost}
\alias{match_FishCost}
\title{Get fisheries cost layer.}
\usage{
match_FishCost(
  PlanUnits,
  group = "all",
  Direc = file.path("~", "SpatPlan_Data")
)
}
\arguments{
\item{PlanUnits}{Planning Units as an \code{sf} object}

\item{group}{The group of fish for the analysis. At the moment the only options are to split by "pelagic" or "all"}

\item{Direc}{The directory where the MME data is being stored. If not specified, the default location is assumed.}
}
\value{
A \code{sf} dataframe of fisheries cost for each planning unit.
}
\description{
Get fisheries cost layer.
}
