% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match_meow.R
\name{match_MEOW}
\alias{match_MEOW}
\title{Assign Spalding's Marine Ecoregions to spatial data}
\usage{
match_MEOW(df)
}
\arguments{
\item{df}{An \code{sf} object to be matched to longhurst}
}
\value{
An \code{sf} dataframe with additional MEOW columns (\code{EcoRegion}, \code{Province}, \code{Realm} and \code{LatZone})
}
\description{
A function that assigns Spalding's Marine Ecoregions and description to each spatial feature.
}
\examples{
out <- utils4mme::PUs_Tasman \%>\%
  match_MEOW()

}
