% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{meow}
\alias{meow}
\title{MEOW}
\format{
\code{meow}
A data frame with XXXX rows and XXX columns:
\describe{
}
}
\source{
"XXXXXX"
}
\usage{
meow
}
\description{
An sf dataframe of
}
\keyword{datasets}
