% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mme_getfishbase.R
\name{mme_getfishbase}
\alias{mme_getfishbase}
\title{Download and return fishbase and sealifebase data}
\usage{
mme_getfishbase(names = NULL)
}
\arguments{
\item{names}{\link{optional} A character vector of species names to subset to}
}
\description{
The function will also subset to \code{names} if wanted but it is recommended you
do this outside the function so you can query the results.
}
\details{
It is essentiall a wrapper around fb_tbl() from the rfishbase package.
}
