% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mme_make_valid.R
\name{mme_make_valid}
\alias{mme_make_valid}
\title{Make sf polygons valid}
\usage{
mme_make_valid(x, group)
}
\arguments{
\item{x}{sf object with invalid geometries}

\item{group}{Group of original data to add the fixed data back into}
}
\value{
a corrected sf object with valid geometries
}
\description{
This function uses an alternative method to try and make sf polygons valid.
sf::st_make_valid often results in incorrect polygons with weird lines.
}
