% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mme_validateTaxonomy.R
\name{mme_validateTaxonomy}
\alias{mme_validateTaxonomy}
\title{Validate taxonomy using World Register of Marine Species (WoRMS)}
\usage{
mme_validateTaxonomy(names)
}
\arguments{
\item{names}{A vector of species (or other) names}
}
\description{
The aim of a World Register of Marine Species (WoRMS) is to provide an
authoritative and comprehensive list of names of marine organisms, including
information on synonymy. While the highest priority goes to valid names,
other names in use are included so that this register can serve as a guide to
interpret taxonomic literature.
}
\details{
The content of WoRMS is controlled by taxonomic and thematic experts, not by
database managers. WoRMS has an editorial management system where each
taxonomic group is represented by an expert who has the authority over the
content, and is responsible for controlling the quality of the information.
Each of these main taxonomic editors can invite several specialists of
smaller groups within their area of responsibility to join them.

This register of marine species grew out of the European Register of Marine
Species (ERMS), and its combination with several other species registers
maintained at the Flanders Marine Institute (VLIZ). Rather than building
separate registers for all projects, and to make sure taxonomy used in these
different projects is consistent, VLIZ developed a consolidated database
called ‘Aphia’. A list of marine species registers included in Aphia is
available here. MarineSpecies.org is the web interface for the marine taxa
available in this Aphia database. WoRMS combines information from Aphia with
other authoritative marine species lists which are maintained by others (e.g.
AlgaeBase, FishBase), the so-called 'externally hosted and managed species
databases'.

Resources to build MarineSpecies.org and Aphia were provided mainly by the EU
Network of Excellence ‘Marine Biodiversity and Ecosystem Functioning’
(MarBEF), and also by the EU funded Species 2000 Europe and ERMS projects.

Aphia contains valid species names, synonyms and vernacular names, and extra
information such as literature and biogeographic data. Besides species names,
Aphia also contains the higher classification in which each scientific name
is linked to its parent taxon. The classification used is a ‘compromise’
between established systems and recent changes. Its aim is to aid data
management, rather than suggest any taxonomic or phylogenetic opinion on
species relationships.

More information can be found at: https://www.marinespecies.org/about.php
}
\examples{
mme_validateTaxonomy(names = c("Jasus edwardsii", "Calanus finmarchicus"))

}
