% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/num2radians.R
\name{num2radians}
\alias{num2radians}
\title{Convert a number to radians (e.g., for day of year or time of day).}
\usage{
num2radians(Number, Max = 365)
}
\arguments{
\item{Number}{A numeric variable}

\item{Max}{A numeric variable for the maximum possible value of the variable (note that this might be larger than in the observations).}
}
\description{
A function that converts a number (e.g., for day of year or time of day) to radians, usually needed before using the harmnonic function in statistical models.
}
\examples{
num2radians(runif(10, 1, 365))
}
\author{
Anthony J. Richardson
}
