% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{PUs_Tasman}
\alias{PUs_Tasman}
\title{Gridded planning units for the Tasman Sea}
\format{
\code{PUs_Tasman}
A data frame with XXXX rows and XXX columns:
\describe{
}
}
\source{
"XXXXXX"
}
\usage{
PUs_Tasman
}
\description{
An sf dataframe of planning units in the Tasman Sea
}
\keyword{datasets}
