% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/untibble.R
\name{untibble}
\alias{untibble}
\title{Remove nonsense attributes if we are working for speed and memory efficiency.}
\usage{
untibble(tibble)
}
\arguments{
\item{tibble}{The tibble to be converted to a data.frame}
}
\value{
\code{df} A stripped down dataframe
}
\description{
Written by Bill Venables (CSIRO/UQ)
}
\examples{
df <- tibble::tibble(x = c(1, 2), y = c("January", "February")) \%>\%
  untibble()
class(df)
}
