% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match_longhurst.R
\name{match_longhurst}
\alias{match_longhurst}
\title{Assign Longhurst Provinces to spatial data}
\usage{
match_longhurst(df)
}
\arguments{
\item{df}{An \code{sf} object to be matched to longhurst}
}
\value{
An \code{sf} dataframe with additional longhurst columns (\code{ProvCode}, \code{ProvDescr})
}
\description{
A function that assigns the Longhurst province and description to each spatial feature.
}
\examples{
df <- data.frame(lon = c(150, 160),
                 lat = c(-32, -38)) \%>\%
  sf::st_as_sf(coords = c("lon", "lat"), crs = "EPSG:4326") \%>\%
  match_longhurst()
}
