% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match_meow.R
\name{match_meow}
\alias{match_meow}
\title{Assign Spalding's Marine Ecoregions to spatial data}
\usage{
match_meow(df)
}
\arguments{
\item{df}{An \code{sf} object to be matched to longhurst}
}
\value{
An \code{sf} dataframe with additional MEOW columns (\code{EcoRegion}, \code{Province}, \code{Realm} and \code{LatZone})
}
\description{
A function that assigns Spalding's Marine Ecoregions and description to each spatial feature.
}
\examples{
df <- data.frame(lon = c(150, 160),
                 lat = c(-32, -38)) \%>\%
  sf::st_as_sf(coords = c("lon", "lat"), crs = "EPSG:4326") \%>\%
  match_meow()
}
