% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/untibble.R
\name{untibble}
\alias{untibble}
\title{Remove nonsense attributes if we are working for speed and memory efficiency.
Written by Bill Venables (CSIRO/UQ)}
\usage{
untibble(tibble)
}
\arguments{
\item{tibble}{The tibble to be converted to a data.frame}
}
\value{
\code{df} A stripped down dataframe
}
\description{
Remove nonsense attributes if we are working for speed and memory efficiency.
Written by Bill Venables (CSIRO/UQ)
}
